requirements.txt
openai
yfinance
python-dotenv
pandas
tqdm
chromadb
matplotlib
seaborn 